requirements.txt
fastapi[standard]==0.115.6
pytest==8.3.4
pillow==11.0.0
PyMuPDF==1.25.1
numpy==2.2.1
